Create README.md
Medication Reminder

Group Hermes Project


Members:
Andrei Karl Garcia
Cedric Santos
Dagne Daphne Pactoranan

BSME-1


PROJECT DESCRIPTION:

A medication reminder is an application that manages to remind or notify the user, when and what medicine should be taken depending on their scheduled medication. The application is capable of setting medication reminders on time in order to notify the user what specific medication and dosage should be taken. It is helpful to users who need an organized medication routine.